+++ src/main/frontend/package.json
@@ -25,6 +25,7 @@
     "write-file-webpack-plugin": "^4.5.1"
   },
   "dependencies": {
+    "@reach/router": "^1.3.3",
     "react": "~16.8.0",
     "react-dom": "~16.8.0",
     "react-router-dom": "^5.1.2",
